This commit was generated by merge_to_master.py.

+++ Source/core/fetch/FontResource.cpp
@@ -31,7 +31,7 @@
 #include "core/fetch/ResourceClientWalker.h"
 #include "core/fetch/TextResourceDecoder.h"
 #include "core/platform/HistogramSupport.h"
-#include "core/platform/SharedBuffer.h"
+#include "platform/SharedBuffer.h"
 #include "core/platform/graphics/FontCustomPlatformData.h"
 #include "core/platform/graphics/FontPlatformData.h"
 #include "wtf/CurrentTime.h"
@@ -100,7 +100,7 @@ FontPlatformData FontResource::platformDataFromCustomData(float size, bool bold,
         return FontPlatformData(size, bold, italic);
 #endif
     ASSERT(m_fontData);
-    return m_fontData->fontPlatformData(static_cast<int>(size), bold, italic, orientation, widthVariant);
+    return m_fontData->fontPlatformData(size, bold, italic, orientation, widthVariant);
 }
 
 #if ENABLE(SVG_FONTS)
